text_jepa_base/requirements.txt
# requirements.txt
torch>=2.0.0
tokenizers>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
torchinfo>=1.8.0
numpy>=1.23.5
